requirements_4_image.txt
apache-airflow-providers-postgres==5.6.0
airflow-clickhouse-plugin==1.1.0
apache-airflow-providers-telegram==4.1.1
plotly==5.17.0
kaleido==0.2.1
pyTelegramBotAPI==4.14.0
beautifulsoup4==4.12.2
fake-useragent==1.3.0
selenium==4.16.0
webdriver-manager==4.0.1